+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 ## Release 3.0.M2
 
+* #177 MutableExtractorModel.getProperties() should respond unmodifiable hash map since 2.9
 * #212 Consistent logging of resource resolving
 * #213 Extractor cache is broken
 * #214 Debug extractor on each execution